Build: handle profile in webpack compilation
Add WEBPACK_MODE as a variable set by different configurations, and use
it when calling webpack.
This is slightly different from "NODE_ENV" because we want to
distinguish 3 use cases:

* Normal usecase: we just want the web compilation to be as fast as
  possible.
* Web development use-case: we want to have the best source maps, and
  use watchers so that changes are immediately taken into account
* Release usecase: optimize the build

This is handled in platform/www/webpack.config.js. The normal usecase is
when mode is left unset. The web dev use-case will be mode=development,
and will be set by a custom script (that also add watchers, etc).

Change-Id: If30f2c01337fdd523909cb5d7cd95a2e68f101c4