waf: Speedup build by using md5_tstamp tool.
By default, waf always hashes the file contents to know if they changed or
not (unlike Make which only considers modification time). It is possible
to change its default behavior by using the md5_tstamp tool: in that case
it evaluates the hashes only if the modification time or size changes).

The official documentation says that the performance impact of using this
mode is "usually insignificant", but this is not what I observe on no-op
builds:

    +------+--------+------------+-----------+
    |      |  Make  | Waf before | Waf after |
    +------+--------+------------+-----------+
    | mmsx | 20 sec | 20 sec     | 5 sec     |
    | one  | 8 sec  | 7 sec      | 0.7 sec   |
    +------+--------+------------+-----------+

(as you can see, waf and Make were equivalent, while waf computed the
hashes of all the repository files!)

This gain makes waf really reactive, which is really really appreciable.

The counter-part is that the build won't be re-made if a file changes in
the same second as the last build occurrence. I don't think this is a
problem (and this was already the case with Make.

Change-Id: If80d20966680a2f3760d1f8cc77caca1a76764a9