waf: Directly build binaries/shared-libraries in the source directory.
By default, waf builds everything, including the binaries/shared-libraries
in the build directory. But we want to have them in the build directory.

To do that, previous code was creating DeployTarget tasks to create
hardlinks of the was outputs in the source directory.

We actually can do something more simple and efficient: directly build
them in the source directory. This is what this patch does.

Change-Id: Ie229710f88ac2d4a247f568a89129d7b58b6acfd