waf: Fix process_use being called twice on some task generators.
This commit reworks the compute_clang_extra_cflags stuff that was manually
calling process_use/propagate_uselib_vars to create the environment of the
Blk2c/Blkk2cc/ClangCheck tasks.

As explained in the removed comment, this was not elegant and was
responsible for some doubled flags in the GCC arguments.

Actually, fixing it was quite easy: just use an "after
propagate_uselib_vars" method.

Change-Id: I7e1f4a53df605cceb2477f636812cfe18b1b185b