waf: Fix idx of .pic task generators.
Problem:
~~~~~~~~
The 'idx' attribute (which is supposed to be unique per path) of the task
generators created for the double fpic were not correct, because they were
actually computed as if the tasks generators belong to the project root
directory.

This lead to possible conflicts in the name of the output files created.

Solution:
~~~~~~~~~
This happened because waf computes the idx according to the path of the
build context when task generators are created. So set it to the wanted
value.