waf: Execute farch and tokens tasks before blk ones.
I observed a build that failed on Igloo with the following error:

  igloo/base.blk:21:10: fatal error: 'iglootokens.h' file not found

This happened because the Blk2c task for preprocessing base.blk was
executed before the Tokens2c task that generates iglootokens.h.

It was possible (but very very unlikely) because there was no constraint
about the execution order.

Change-Id: Iabbc95eaa237ea52d7950ddc2fba248443ac60dc