waf[common]: Execute CustomInstall tasks after vnum ones.
In the tcs-hpe directory of the tools repository, the installation task
was executed before the vnum one, that creates the library symlink that is
needed for installation, leading to installation failures.

This happened because apply_vnum method is called after add_custom_install
one.

Fix that by using the 'after' attribute of the CustomInstall task.

Change-Id: I58cbe8b56b0a2b2013eb8626b289968e8210696e