Make: Fix "make check-translations" not generating tpl files.
Problem:
~~~~~~~~
The "make check-translations" command was not generating the translate.tpl
files of the products. When a patch is built on the buildbot, they were
not generated at all, because the buildbot does not run "make i18n".

This was not an issue before because we do not have selenim tests about
translations. *BUT* this became a problem since the recent patches of
Benoit in One about the single-page translations in One, because now this
is the javascript that fetches the tpl files, and it caused errors in the
browser.

Solution:
~~~~~~~~~
Call "make i18n" from one/www in the translations target. This target runs
"make xgettext locales" from www/po (with locales including merge), plus
generates the One's translate.tpl (we don't care about other products).

This is really dirty, but I can't see another easy way of fixing it, and
we'll probably get rid of Make (waf does not have the issue) so...

Change-Id: Idab01d773709ee4f9f49e3de80e73c6ee9dddfd8