iop-dso: fix opening same iop dso multiple times.
When opening the same iop dso multiple times, dlmopen will return the
same handler each time, and so the pointers to the packages and types
are the same.
Since we compare the pointers of the packages and types to see if they
can be register and unregistered when loading and unloading the dso, we
have some conflicts when the same dso is opened multiple times.
To fix it, we should have a 1:1 relation between the handler and the iop
dso.
So when we detect that the handle as already been opened, dup the
previous iop dso instead of creating a new one.

Change-Id: I2bd37342c5f6592ef2c27fcf17e2312bf1efcee1