waf/Build: add eslint as javascript ale linter
This linter will be used for javascript files in master.
To do this, we update the .local_vim.vim file with a new ale_linters
configuration.
This configuration will override the last one defined in .vimrc for
javascript files only.
Thus:
* before this patch, JS files are linted with JSHint,
* after this patch, JS files are linted with eslint

Note: `make ale` needs to be called to refresh the .local_vim.vim, or `waf
configure`

Change-Id: I899b7e30c92f2e37b8569ab4e848c4e84bfd0aa2