net/sctp: provide a new helper to use the latest CONNECTX3
The SCTP_SOCKOPT_CONNECTX3 allows to initiate a SCTP connection and to
retrieve the new association ID immediately. This API is mandatory to
correctly close a hanging connection in the one-to-many use cases (thus
to have a clean fix for #63116).

The new sctp_connectx_ng() will fallback on the old CONNECTX API if the
association ID is not required.

Note that if this API is part of the kernel for a while (2.6.31 which is
available with Redhat 6), its support is part of libsctp since the
1.0.12 which is only available in Redhat 7... Thus we have to hardcode
the value of SCTP_SOCKOPT_CONNECTX3 if we want to use it with Redhat 6.

Refs: #63116
Change-Id: I7f9a92bb261258856676b9378fa287bcc460bb6f