log: Fix va_list not copied in logger_putv.
"va" was used without being copied and reused afterwards, which was
obviously buggy. Because of that, using "log_start_buffering(true)" was
dangerous.

But because we are "lucky", it seems to have a real bad consequence only
if we use a special formater (%*pS or equivalent) and with a non-silent
logger... but not always!

It was responsible for a crash of the mde tests in mmsx, but for a still
unknown reason it crashed only in the following conditions:
* built with waf.
* on a buildslave.
* run with `zchk mde` or `zchk mde.trace` (but not with `zchk` nor
  `waf check`), OR when the buildbot runs `waf check`.

I don't even want to know why... The test added reproduces the crash (at
least on my machine, using waf).

Change-Id: I55479c64fa1099dc01b5ee8a56317ea036334082