iop: do not end json with newline when JPACK_MINIMAL is used
the IOP_JPACK_MINIMAL flag is described as trying to produce the
smallest json possible, but it always included a trailing newline
character. That meant that in the codebase, the flag was almost always
used with the "IOP_JPACK_NO_TRAILING_EOL" one, to remove this newline
character.

The very few occurrences of MINIMAL without NO_TRAILING_EOL do not
depend on the inclusion of this character, so it makes much more sense
to include NO_TRAILING_EOL by default in this MINIMAL flag.

Change-Id: I5bdee62184ffe43f325ebb94daa43bcefe4978b6