waf: Remove -P option at configure.
This patch is necessary for the following one, in which we'll have a
different build directory for each compilation profile.

This is something that can't be done without getting rid of this option,
because when the 'out' variable is evaluated in the main wscript, the
options are not processed yet (and thus we can't know which is the
selected profile).

So get rid of this option, which was not used in scripts anyways, and that
can be replaced by the P environment variable.

Change-Id: I5ad9ee1e928bc2ce30bd4255fc1c99844e3a43d1