waf: Add FAKE_VERSIONS mode.
When this variable is set at configure, the version files are generated
with fake (and constant) data, so that changing of git revision does not
trigger a re-link of all the binaries.

This is a huge gain of time, but it's not possible to know the revision of
the binaries that are built with this flag, which can be useful for debug.

This is why it's forbidden with release profile.

Change-Id: Ia6826779b42d4eb45a41a886d94bb15356919914