qhash: make sure for_each macros are scope safe without parentheses.
== Problem ==

Before, the for each macros for qh and qm were not scope safe without
parentheses.
The following code led to a wrong result:

    for (int i = 0; i < sl; i++)
        qh_for_each_pos(u32, pos, qh)
            sum += qh->keys[pos];

This was due to the static qhash type check.

== Solution ==

Instead of doing the static type check before the loop, do it in the
initiation step of the loop.

Despite this kind of constructions are against our coding rules, we want
our 'for_each' macros to be as close as a standard 'for' as possible,
and that we may need it to build other macros.

Change-Id: I4e892b29f7218ef547b74130310c5ac9966cd6a1
Refs: #58860