log: fix tracing with logger_t objects
Because of 6a96308447bbe02756d1a9bc89e4cbf042c022aa the tracing were
broken for all logger_t objects ie. (logger_log and logger_trace APIs).

logger_log() used to call internally logger_is_traced() which was very
wrong, because logger_is_traced() is linked to the caller context (file
+ function name), and the result of logger_is_traced() is cached in
a static variable.

So the right way to have logger_log() check the tracing level correctly,
is to call logger_is_traced in the __LOGGER_LOG macro which will be in
the right context. It also allows to get rid of the __LOGGER_TRACE macro
which is now useless.

Of course calling directly __logger_log() is now broken, but it was
maybe less broken before but still broken, the logger_log() function
must be used instead of __logger_log().

Change-Id: I0110d35323aa7b996339fcfe22bc248333ac7f96