iop-json: Use 4 spaces instead of tab for indentation.
Since the beginning, the IOP json packer uses tabs, and not 4 spaces as
specified in our coding rules, for indentation.

The original reason was probably to save some space, especially when it
comes to network. But I do think this was a bad reason because network
layers using json are using compact json anyway, without indentation.

But is was really annoying when editing json files produced by our packer,
because our editors are configured to highlight tabs so that we don't use
them (well, in depends on the situation, some probably don't show the
difference). Worse, in many cases we were committing configuration files
mixing tags and spaces (especially the people having editors that do not
show the difference).

This commit turns indentation to 4 spaces, which should be more convenient
for us, without impact o the network.

Change-Id: I4cd93211b3663991dfe704b04c45f64e931409e4