iop-xml-wsdl: Fix namespacing of enumerations
Enumerations were not namespaced when added in the WSDL file. This
is a big issue, as two different IOP files can declare an enum with the
same name, and it will conflict in the WSDL.
This is what happened in #65327.

Fix this by implementing the TODOs left in the code: use fullname and
not "name". I have no idea why those TODOs were left in place, when
every other type used fullname already.

Closes: #65327
Change-Id: Ic1778fb95633a3ac013df9fd079fbb12a3b89aa5