z: improve string print with newlines on errors
When a Z_ASSERT_STREQUAL is used, if the test fails, both strings
are normally printed, like this:

a (len:3):
  foo
b (len:3):
  bar

However, if a string contains a newline, this is not considered
printable, so we end up with:

a (len:...):
  X'ABE5189C546'
...

This makes it very hard to compare, especially when the string is big.

This patch just considers that whitespaces can be printed out too.

We end up with something like:

a (len:7):
  foo\nbar
b (len:7):
  foo\nbaz

Change-Id: If964c46424d0a5aa2503c5d48fa5da558ea390a0