waf: remove JAVA configuration & compilation tasks
Cherry picked from commit e7ea80134a39b26a1d3f634797f1f5158c3896d7.

Java was only used in one project that uses lib-common and not in
lib-common itself.
We should not check for Java if we only want to use lib-common, we
should do it in the project that uses it.
We also have since removed Java support from that project so it is
useless now.

Change-Id: Icfecca3cf3a2d08efb406210c83711ea31bea697