Introduce ifork() / add child pid in at_fork_on_parent method.
The purpose of this commit is to add the pid of the child process in the
at_fork_on_parent module method.

Since this information is not given by the hook of pthread_atfork, we
cannot rely on it to get this information. This is why this commit
introduces ifork() as a wrapper around fork(), which calls the methods
itself. fork() is now forbidden, to ensure the module method can be
properly called.

But there are still some situations where we do not control the fork()
calls (like in python modules for example), so we have to keep the
pthread_atfork handlers. In these cases, the child pid is unknown when
calling the method.

Change-Id: I2a42c1d00d07ecf29ca102ec396c456e29d87025
Refs: #65362