el: Register el_sigchld_hook at initialization.
Since previous commit, we have to be sure el_sigchld_hook is called even
if el_child_register was not called for the fork process.

So register it at initialize instead of doing it lazily (I guess it was
done like that because, a long time ago, there was no module for el code).

Change-Id: I8a04d23ff1c5e09986e5915ee025a68f344235e0
Refs: #65362