iop: support filtering/sorting on an array length with '.len' attribute
In 'iop_msort' and 'iop_filter', allow the use of '.len' attribute for
array length.

Before the change Ie51a7700e74fc5fee7dffe204c9da7face32de23, 'iop_msort'
supported sorting on array fields by using the array length. The change
enabled sorting on full array content instead. This patch reestablishes
the feature but we have to state explicitly in the path that we want to
sort the fields depending on length.

Refs: #59237
Change-Id: Ibe839f9889c3410ef48e387f6b5bb4c2bf12f03b