log: fix caching of the tracing level in P=debug
Since the introduction of the logger API the mechanism to cache the
tracing level was broken in P=debug (which have been seen thanks to the
zchk log.log_tracing test added recently).

The tracing level caching, relies on __builtin_constant_p() to detect
whether the given level is constant or not, and then if it is actually
constant, stores the result of __logger_is_traced() in a static
variable.

However, the tracing level is first pass to __LOGGER_LOG(), which caches
it in a local const variable to avoid multi-evaluations, then passes it
to __LOGGER_HAS_LEVEL() and finally to logger_is_traced().

In O2, GCC is able to get rid of the intermediate const variable
thanks to the inlining and correctly sees if the provided level is
constant. In O0, the const variable is preserved and thus
__builtin_constant_p() always returns false.

So, to keep both the multi-evaluation protection in the non-constant
case, and the tracing level caching in the constant case, we split the
__LOGGER_LOG() code in two parts. One to handle the constant case and
preserved the original parameter until we call __LOGGER_HAS_LEVEL(). And
the other part to just handle multi-evaluations.

Change-Id: I9420de4ca274e74e14666596abe9f3063762bf57