iop: enable array indexes in field paths for filtering and sorting
Before this patch, sorting/filtering paths could not go through arrays
like in the following examples:

   'foo.bar[2]'
   'tab[0].v'

This patch enables this type of use cases that *might* be useful in some
corner case.

This step is mostly a preparatory step for #59237, for which we want to be
able to list the values for a field path containing wildcards:

   'elements[*].name'

Doing that feature without supporting constant indexes would have seemed
odd. Introducing constant indexes in field paths also decreases the amount
of needed work for wildcards, thus making the next patch to come for
feature #59237 easier to review.

Refs: #59237
Change-Id: Ic3f5d0f6dd33ff67137d6a7283ac02772f2c4323