iopc: fix handling of UTF8 BOM
see https://en.wikipedia.org/wiki/Byte_order_mark

The representation of the UTF8 BOM is the following:
Octal | Hexadecimal | Decimal
357   | EF          | 239
273   | BB          | 187
277   | BF          | 191

So this was clearly a bad mix of decimal and octal representations.
\357\273\277 would have been correct as well, it means the same thing in C
strings and in flex.

The bug dates from the first version of iopc :
7bc59f946839f5578f6110104850bacaf4c4c52f
and has been copy/pasted to other lexers.

To test, you can for example open a IOP file in vim.
To save the file with a BOM:
:setlocal bomb
:w
To ask Vim if there is a BOM:
:setlocal bomb?

Also possible with echo -e '\xEF\xBB\xBF' and hexdump -C.

Change-Id: Ie002aac0af2a1c799d661ff70fab0652832372aa