iop: use 'iop_cmp_field' to improve implementation of 'iop_sort'
'iop_cmp_field' is a recent function that defines an arbitrary order
between two IOP fields values (even repeated or constructed).

'iop_sort'/'iop_msort' did quite a similar job for sorting IOP structures
depending on a field but it was using a different code and didn't support
sorting on repeated or constructed fields except in some marginal cases.

This patch adds the function 'iop_cmp_value' and uses it to unlock sorting
based on repeated or constructed fields.

The code of 'iop_cmp_value' contains the original code of 'iop_cmp_field'
but without taking a 'field' argument: that lets us manage the need for
pointer indirection by ourselves (the pointers coming from
'iop_get_fieldp' rarely need an extra indirection).

The behavior when sorting IOPs with a repeated field is changed because we
now compare their content when we used to compare their array lengths.
This way of comparing repeated fields is going to be reestablished in a
subsequent patch.

Done as preparatory work for #59237: I preferred applying this change
before extending the possibilities of IOP field paths any further.

Refs: #59237
Change-Id: Ie51a7700e74fc5fee7dffe204c9da7face32de23