iop: rethink functions 'iop_value_has' and 'iop_opt_field_is_set'
iop_value_has()
~~~~~~~~~~~~~~~
The naming made absolutely no sense: the function only applied on optional
fields. Renamed it to 'iop_opt_field_isset()'.

The field description passed to the function was almost unused: we only
used the type, so the prototype of the function was updated so we pass
only the type of the field.

iop_opt_field_is_set()
~~~~~~~~~~~~~~~~~~~~~~
Became quite redundant with 'iop_opt_field_isset()' after the changes
listed above but I didn't want to merge the function to avoid any impact
on performance.

This function becomes 'iop_opt_field_getv()', and now returns -1 if the
field is not set (instead of returning a boolean).

Function 'iop_opt_field_is_set_const()' was updated accordingly.

Seen while working on #59237.

Change-Id: I876284df4604ab0c0096a770e95ffc2cbd80a709
Refs: #59237