Revert "el: Register el_sigchld_hook at initialization."
This reverts commit e0fdbf33f0d4379870e17425eac01275bf1ac39d.

A consequence of this patch is that we always call waitpid for all the
forks (included the ones that were created by an external library), as
soon as the el module is required.

Before this patch, it was only the case if el_child_register was called at
least once.

So it changed the behavior of iopy, which does not call el_child_register
at all. This was responsible for making the iopy tests fail:

 https://git.corp/buildbot/builders/review-tools_waf/builds/38/steps/test-waf-check/logs/report

Change-Id: Ifa8e1974dbfb50ca5a3559f6eead351228f3d720