iop: refactor field parsing between two functions
Use a single function to parse fields for 'iop_get_field_const' and
't_iop_compile_field_path'.

Also enable the use of negative indexes for 'iop_get_field_const'. In this
case, the arrays are read backward: 'foo[-1]' is the last element of array
'foo'.

Note: we may also want to homogenize the behaviors for '_class' attribute,
returning the full name of a class instance and 'len', returning the
length of an array.

Change-Id: I0bfed95d0e03286de000791e58e356a69bf9d6c7
Refs: #59237