check_php: do not iterate on submodules
The check_php.py is a script that is used to verify that php files
are compatible with the buildbot's version of php.

It iterates not only on every php files in subdirectories, but also on
every submodule.
This is done with weird git commands, that actually returns every
submodule that has existed in the repository, not just the current
submodules.

All of this is simplified to just iterate on every php files found in
subdirectories, without using git. Yes, it thus checks php files that
are not tracked in git, but this not a situation that can really
happen.

This allows simplifying the code quite a bit: we now properly print
which files are checked (instead of just saying "Check www"). And we
no longer depend on ipath.

Change-Id: Ibab755320f22fe722eef3e2c079d6a6b812981c4