log: fix wiping parent logger used in thr log test.
The parent logger used in thr log test was not wiped at the end of the
scope.
So when __logger_do_refresh() was called in another test, the root
logger still had a reference to the thr test parent logger leading to
stack-use-after-scope.
The solution is to simply wipe the parent logger at the end of the thr
test.

Change-Id: I6fe36d68757a2b2341cb4021c10211ed974265ae
Closes: #65518