iop-rpc-channel: fix _ic_read errno handling and return type.
When a socket has been gracefully closed, read returns 0 and let errno
unchanged. Thus, when errno is EAGAIN at the start of _ic_read(), that
code may let the caller that an incomplete read is going on. This is a
lie.

When it happens via ic_version_handshake(), the socket will always be
considered as available to read, and ic_version_handshake() will be
called forever by the event loop, thus leading to a 100% CPU.

I don't remmeber why I did this code but it's clearly a nonsense now
because both sb_read and ssl_sb_read behaves as read and recvmsg:
  * -1 on error — then errno is set with either EAGAIN or EWOULDBLOCK,
  * 0 on graceful exit (EOF) — errno is not set,
  * > 0 on success — errno is not set.

In this patch, I go back to the previous behaviour, which was correct,
i.e. returning -1 on graceful exit and setting errno to 0 to avoid
that underlying functions believe that the call is retriable. I also
use a more consistent behaviour of handshake functions and describe
the new behaviour. This allow us to get rid of checking if the IC is
queuable to know if the connexion succeed. Instead, we just check if
there was an error.

Also fix the return type of _ic_read(), it should be a ssize_t instead
of an int.

Change-Id: I95c908dc207611b723f67a4c75e069bbb1be2191
Closes: #62195