qpages: fix odd condition in free_n
The function `free_n` has conditions to call `madvise` with MADV_DONTNEED
that enable to release unused memory.

The first condition checks if the block size (bsz) is the same as the
number of pages for the given page_run_t*. This means the whole arena is
free and we can madvise that we do not need this arena.

The second condition checks if we free more than QDB_MADVISE_THRESHOLD
(i.e. 256) pages. In this case we call madvise on a chunk of allocated
memory by the arena. This second condition seems intended for QDB.

QDB will allocate/free chunks of memory for its columns using
qpage_alloc* and qpage_free* functions. But it will only allocate/free one
page per chunk, so we never enter the second condition …

When freeing pages we compute chunks of QDB_MADVISE_THRESHOLD for the
given page_run_t*. When freeing the last page of a chunk we can call
madvise on it.

Refs: #63021
Change-Id: I184ae4a7cb158fc39cc18f11a9b1e07c6a3053e0