el: rework SIGCHLD handler and register it at initialization.
Following e0fdbf33f, we want to register the callback for the SIGCHLD
signal at startup and not lazily.

The issue, as seen here 28410f451f, is that before, in el_sigchld_hook()
we called waitpid() for all the child processes, with waitpid(-1, ...).

This is an issue when we don't actually control the fork() done by
external libraries (for example in python with IOPy).

We only want to call waitpid() for the child processes that are tracked
by the el library and let the external libraries do their own waitpid().

Before, only the child processes registered with el_child_register()
were tracked by the el library.

In our case, we also want to track the processes created with ifork().
So let's use the at_fork_on_parent module method to register the
processes with a NULL el, so we can call at_fork_on_child_terminated
even it that case.

Change-Id: Icd1c7fa1b54a01dd41c49d101c5a8f077323546e
Refs: #65362