iop: simplify 'iop_opt_field_getv()' prototype
The result of this function was split between the return value (-1 in case
of absent field) and the pointer on the value (given as a 'void **'
parameter).

The reason it was split was simply that we didn't really know what to do
with the optional void case: what pointer value should we return in that
case ?

In fact, this is a false problem: the function user isn't supposed to use
the pointer value without considering its type, so it's not supposed to do
anything with a pointer on a optional void field, except checking whether
it's null or not, so any 'slop' pointer value is supposed to work.

This should slightly simplify the uses of the function and also help us
getting rid of smelly constructions like:

    res = iop_opt_field_getv(type, data, &data);

Change-Id: I987187d50802fa7d72531d88318f19ffb879927b