qps: avoid ld warning with asan
/usr/bin/ld: warning: Cannot export local symbol '__asan_default_options'

This warning appears on my machine but also on centos 6 bots (at least).

The solution of setting the default visibility for __asan_default_options
was found here :
https://bugs.chromium.org/p/chromium/issues/detail?id=368351#c9

Setting allow_user_segv_handler=1 is no longer needed (and has been
deprecated) starting with clang 5, so remove it in that case.

I checked that ./zchk qps still works in asan with both clang 3.9 and
clang 7.

Change-Id: I34f943379911f2a097c4d841a8ccb16c283728f3