file-bin: fix file_bin_get_entry_end_off()
If the last record of a file ends at the end of a slot, the function
file_bin_get_entry_end_off() does not return the offset at the end of
the record but the value plus the size of a slot header (because the
function get_multiples_nb_in_range() includes its boundaries in the
computation).
This patch aims to fix that by rewriting this function without using
get_multiples_nb_in_range().
file_bin_get_next_entry_off() also needs to be rewritten because fixing
the bug in the previous function creates another bug here (this function
was actually compensating the bug).

Change-Id: I6f7966eebcf34070164787759da8de4f27f19981
Closes: #62955