core/module: rework argument providing in 'MODULE_BEGIN' blocks
Keep the argument with the requiring module instead of keeping it with the
to-be-required one.

In the patch 5c7024cb9f0c8dcd10c2822a0ae8bdb85f2a103b (2014), we added the
support of arguments providing into module definition. That was useful to
get neat module definition with module arguments provided along with
dependencies, but there was a problem along the way.

Context:
~~~~~~~~
The 'MODULE_BEGIN' macro defines a function that registers the module. The
function has the 'constructor' attribute so it is run early when the
program runs.

Problem:
~~~~~~~~
The argument is stored along with the to-be-required module, so the
following case can happen:

   ------ foo.c -------
   MODULE_BEGIN(foo)
      static bar_cfg_t bar_cfg;
      MODULE_PROVIDE(bar, &bar_cfg);
      MODULE_DEPENDS_ON(bar);
   MODULE_END()

   ------ zchk.c ------
   Z_GROUP_EXPORT(bar) {
      MODULE_REQUIRE(bar);

      /* Tests. */

      MODULE_RELEASE(bar);
   } Z_GROUP_END;

In zchk.c, we'd expect 'bar_initialize()' to be called with a null
argument because we didn't provide any argument inside the Z_GROUP. But if
zchk.c was built along with 'foo.c', we're going to get the argument
'bar_cfg' instead, even if we never required the 'foo' module.

Solution:
~~~~~~~~~
Keep the arguments provided inside 'foo' definition in a new map
'deps_arg' inside 'foo' module. Then, when loading 'bar' as a dependency
of foo, retrieve and use the argument.

Note1: if an argument is already provided manually, don't apply the
argument previously provided in the definition block so we can stick to
the current behavior (the argument provided in the definition block was
overridden).

Note2: the previous map 'modules_arg' doesn't need to be kept because it
was specifically designed for providing arguments to modules that are not
yet registered, which is not usually expected to happen outside
'MODULE_BEGIN' blocks.

Change-Id: Id4807d5167c0e3d7f5e211ef6b53bfd242f2653c