ic: fix TLS handshake on blocking socket having a timeout.
Here is a tricky bug: consider you called ic_connect_blocking with two
arguments (omiting the timeout argument). I bet we have no clue what
could be in there but it's likely not to be 0. And here is that we
observed OpenSSL firering an SSL_ERROR_WANT_READ (while the socket is
actually blocking). I suspect that it's because we reach the
timeout. In this case, we must close the IC and return an error.

If timeout on blocking socket is actually the root of our problem,
it's not surprising we didn't observed it repeatitively: the network
has been fast enough to do the connection, to receive the version
handshake, but failed here with the TLS handshake.

Bug found with: I89f6442c5cafddd78af73c0833deda5275acdd6b

Refs: #64827
Change-Id: Ia019afbccd8f8ad23abc9a4f25e7040ad989545d