core/module: document and use parameter 'module' in 'module_register'
The parameter 'module' used in 'module_register' contains a static pointer
to the module that can be used to prepare arguments or dependencies for
modules that aren't registered yet.

Document the parameter and set the pointer during the call to
'module_register' instead of doing into the macro 'MODULE_BEGIN'.

Change-Id: I4f0afaf45fdcce30972e1deb2cbb5748696ac2d4