core/module: rework in-initialization module argument providing
Make module argument providing during initialization more consistent with
regular argument providing.

The patch b972f3c17876ec3d587deb43664ef9c7b1c2c639 introduces a new map to
save the argument provided by each module to the current ones. The issue
the patch is trying to fix is that the argument was provided even if the
module providing it wasn't actually loaded at all.

This patch makes it a little more closer to how argument providing
actually works: it simply delays the provision of arguments for other
modules until the module is actually loaded.

In order to maximize the consistency with the previous behavior, the
provided argument don't override the arguments that are provided out of
module initialization: they used to be provisioned first and overridden
afterwards so they should not override any other argument now that it is
more likely that they are provided later during module loading.

This patch also fixes the following case:

 A depends on B
 B depends on C
 A provides an argument to C during initialization

With the previous implementation, the argument wasn't transmitted to C
because there A didn't have a direct dependency on C.

Change-Id: I34337bd6fe8c157338d3255da4b86d2ce122c3df