iopc/ts: improve typings when registering collections of classes
The registerCollection function is generic on the model of the
collection.
This works fine for structs, because Typescript properly infers the
model from the collection description. However, this no longer
typechecks with Typescript 3.5 for classes, as Typescript fails to
deduce the generic and reverts to the base "IopModel".

To fix this, explicit the generic when calling the function.

Change-Id: I1a62d3a44206b839632474530361a01f5c6a0ef8