waf: Fix double fPIC compil in case of a library using another one.
The previous patches, we now have the 'libcommon' library that uses the
'libcommon-iop' and 'libcommon-minimal' libraries. The binaries/shared
libs that need the lib-common actually use 'libcommon' only, and the other
2 are infered by waf.

This is great, but did not work fine with the double fPIC compilation,
because only the libraries directly used by the shared libraries task
generators were considered. Because of that, some shared libraries were
linked against non-fPIC static library, which does not work.

To fix this issue, simply recurse in process_use_pic.

Change-Id: Iea66a8a59a61ce2d06ed06bcff16e1cfcd35b21e
(cherry picked from commit bbdee7f1ac760366613329edc0452420721ece06)