dblatex: fix regression in PDF header on debian9
In LaTeX, sections can be different according to the document class
being used:
* for "book", the topmost section is "chapter", then "section", then
  "subsection", ...
* for "article", the topmost section is "section", then "subsection",
  ... DocBooks are generated as "article" class.

In the header, we chose to put "\leftmark", which is defined as
containing the name of the current "chapter". It is very unclear in
LaTeX doc whether this literally means "chapter", so useless for
"articles", or rather means a more loose "topmost section".

Anyway, in debian8, the header was containing the section title. In
debian9, however, the header was only containing "CONTENTS", that is,
the title for the table of contents.

This bug is probably due to a buggy .tex file generated by dblatex,
causing issues with the "fancyhdr" package which redefines the macros
used to mark the chapter/sections. I say probably because this is an
absolute nightmare to debug, with interlaced latex packages each causing
issues with the others. The bug could be in dblatex, in pdftex, in
fancyhdr, in another package, I have no idea.

Normally, a chapter with mark both "left" and "right" marks,
but a section will mark only "right" marks (so, \leftmark is current
chapter, and \rightmark is current section. Please note the complete
confusion about how this applies on "article" classes, where there is no
chapters...).
Using \rightmark solves the issue on debian9, but causes builds on
debian8 to use subsections...
The solution picked is to redefine the function called for each section,
to define both the left and right marks. that way, \leftmark will use
section titles on debian8 and debian9.

As a side note, this is exactly why we want to get rid of
asciidoc+dblatex. LaTeX is already very complex to use when the tex file
is written by hand, but having several tools generated tex files using
multiple packages in order to generate a PDF file means this sort of
ugly bugs.

Change-Id: Ida9ac0b195d875167f8399652bba330b6125213e