iopy: rework iopy ic server mutex callback handling.
When calling the iopy ic server callbacks, we want to release the el
mutex before locking the GIL, so the two locks don't conflict.
This has the side effect that the iopy ic server can be deleted while
being in a callback.
To solve this issue, the iopy ic server is now protected with a
reference counter and the python context is now protected with the GIL.

The server is new also always wiped by the event loop.

Change-Id: Iaed70fb3f71a885be0286d36b0ab5c09e7fd4e22