iop: introduce 'iop_struct_for_each_field'
What ?
~~~~~~
A macro that iterates through the fields of either a class or a simple
struct/union.

Why ?
~~~~~
In order to do that, there are many places in the code in which we do the
following:

    if (iop_struct_is_class(st)) {
        const iop_field_t *fdesc;
        const iop_struct_t *cls;

        iop_class_for_each_field(fdesc, csl, st) {
            do_things(cls, fdesc);
        }
    } else {
        for (int i = 0; i < st->fields_len; i++) {
            do_things(st, &st->fields[i]);
        }
    }

The code above is correct and is (for now) the right way to iterate
through fields with the macros and functions we already have.

There are also some cases in the code where we should have written the
code like in the example above but did this instead:

    for (int i = 0; i < st->fields_len; i++) {
        do_things(st, &st->fields[i]);
    }

Of course, this is broken if the struct is a class: the parent classes
fields are going to be ignored while they potentially don't ought to be.

Now it could be written this way:

    iop_struct_for_each_field(fdesc, _st, st) {
        do_things(_st, fdesc);
    }

Note1: for case of simple structs, 'iop_struct_for_each_field' is around
2.5x slower than the basic 'for' loops so it can be considered when
performance is a major concern.

Note2: bench tool (added in a subsequent patch) has shown that the new
macro can be faster (~10%) than the old one when used for classes.

Note3: tests are going to be added in a subsequent patch, when
'iop_class_for_each_field' is going to be removed.

Change-Id: Ifa5f64c6719efd9b0c2bb379f4de5b7a32408697