bench: move most of the identified bench material in a bench directory
Bench code shouldn't be mixed with lib-common library code but it doesn't
means that it should be thrown away after use.

This new directory is meant to contain all bench material created for
lib-common library with the purpose to share and compare benching
techniques in order to erect some "benching good practice".

Change-Id: Iab1845388aec955c257b727928b61756b2a645bf