iopy: rework iopy ic client mutex callback handling.
Same as the server, we want to release the el mutex before calling
iopy_ic_py_client_on_disconnect().

The solution is similar to the server, we protect the python context
with the GIL.

But, unfortunately, we cannot use a reference counter because we cannot
wipe the ichannel when we are in the on_event callback, which can happen
now since we are releasing the el mutex in iopy_ic_client_on_event().

The solution is not to delete the ic_client on iopy_ic_client_destroy(),
but wait for the el thread to do it. This way, we are sure we are not in
the on_event callback.

Change-Id: I69e412609b70c4b6c395f268eb3dbd1724ce868a