iopc: fix assert with invalid @snmpParamsFrom
Spotted with AFL.

When @snmpParamsFrom refers to a snmpObj from another package, there was
not any check when building the package path, the path could be totally
empty, or with empty components, and this was triggering an assert in
check_path_exists -> pretty_path_dot -> sb_shrink.

Also fix leaking of path_new when check_path_exists fails.

Change-Id: Ic731355ac421839592303dfcbae696d6f768a140